refactor: split strore files
