+++ functions/src/index.ts
@@ -3,17 +3,12 @@ import * as functions from "firebase-functions";
 import { cronHandler } from "./cron";
 import { INTERVAL_MINUTES } from "./models/schedule";
 import { createSlackApp } from "./slack";
-import { InstallationStore, RotationStore } from "./store";
+import { createStores } from "./stores";
 
 admin.initializeApp();
 
-const db = admin.firestore();
-const rotationStore = new RotationStore(db);
-const installationStore = new InstallationStore(db);
-const { slackHandler, postRotation } = createSlackApp(
-  rotationStore,
-  installationStore
-);
+const stores = createStores(admin.firestore());
+const { slackHandler, postRotation } = createSlackApp(stores);
 
 const functionBuilder = functions.region("asia-northeast1");
 
@@ -22,4 +17,4 @@ export const slack = functionBuilder.https.onRequest(slackHandler);
 export const cron = functionBuilder.pubsub
   .schedule(`every ${INTERVAL_MINUTES} minutes synchronized`)
   .timeZone("Asia/Tokyo")
-  .onRun(cronHandler(rotationStore, postRotation));
+  .onRun(cronHandler(stores.rotationStore, postRotation));
